[Multiwindowize] main: Rewrite main method for multiwindow support
Before, DesktopNaotu was a pure single-window program. It only
maintained a MainWindow instance.

To make it support multiwindow, we need to create a Set() object to
maintain a window list, and have an independent CreateWindow() method.

Now the program can run as before. However, this is
just a start. There are still a lot to do,
especially that we need to handle front-end global objects
(Angular, Bootbox, Kityminder etc.) properly.

Reference:
- https://www.cnblogs.com/sanshengshui/p/11334568.html